% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{cMiss}
\alias{cMiss}
\title{For a data frame, returns per column the proportion of missing values}
\usage{
cMiss(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a vector of proportions
}
\description{
For a data frame, returns per column the proportion of missing values
}
