% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{cUnique}
\alias{cUnique}
\title{For a data frame, returns per column the number of unique values}
\usage{
cUnique(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a vector of proportions
}
\description{
For a data frame, returns per column the number of unique values
}
