% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{f2n}
\alias{f2n}
\title{Convert factor into numeric}
\usage{
f2n(f)
}
\arguments{
\item{f}{a factor that contains values that can be converted with as.numeric}
}
\description{
Convert factor into numeric
}
