% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{headn}
\alias{headn}
\title{Array-based version of `head` (default = 6)}
\usage{
headn(x, ...)
}
\arguments{
\item{x}{an array (vector, matrix, etc.) or data frame}

\item{...}{a scalar or vector. If scalar, all dimensions will have the same number of elements. If vector, it should match the length of the number of dimensions.}
}
\description{
Array-based version of `head` (default = 6)
}
