% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{Mode}
\alias{Mode}
\title{Finds the modes (modi) central tendency}
\usage{
Mode(x, out = NULL, warn = TRUE)
}
\arguments{
\item{x}{a vector/factor for which to find the mode}

\item{out}{an integer that defines the maximum number of modi to display}

\item{warn}{a logical; if TRUE, a warning is displayed if more modi are found than `out`}
}
\description{
Finds the modes (modi) central tendency
}
