% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{pMax}
\alias{pMax}
\title{For a data frame, returns per column the maximal proportion of values that are the same}
\usage{
pMax(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a vector of proportions
}
\description{
For a data frame, returns per column the maximal proportion of values that are the same
}
