% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{rMiss}
\alias{rMiss}
\title{For a data frame, returns per row the proportion of missing values}
\usage{
rMiss(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a vector of proportions
}
\description{
For a data frame, returns per row the proportion of missing values
}
