\name{SLmisc-package}
\alias{SLmisc-package}
\alias{SLmisc}
\docType{package}
\title{
Miscellaneous Functions for analysis of gene expression data at SIRS-Lab GmbH
}
\description{
Miscellaneous Functions for analysis of gene expression data at SIRS-Lab GmbH
}
\details{
\tabular{ll}{
Package: \tab SLmisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-01-02\cr
Depends: \tab R(>= 2.4.0), stats, graphics, marray, limma, plotrix(>= 2.1.6)\cr
Suggests: gplots
License: \tab GPL (version 2 or later)\cr
}
require(SLmisc)
}
\author{ Dr. Matthias Kohl (SIRS-Lab GmbH) \url{http://www.sirs-lab.com}

Maintainer: Dr. Matthias Kohl  \email{kohl@sirs-lab.com}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
