% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{tableNA}
\alias{tableNA}
\title{Wrapper for table(..., useNA = "always")}
\usage{
tableNA(...)
}
\arguments{
\item{...}{Further arguments for `table`}
}
\description{
Wrapper for table(..., useNA = "always")
}
