% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_table.R
\name{desc_table}
\alias{desc_table}
\title{Create descriptive table by group
Current version works in the html setting. We are working on extend it to the word document}
\usage{
desc_table(varlist, data, groupvar = NULL, total = TRUE)
}
\arguments{
\item{varlist}{One-dimensional predictor}

\item{data}{Dataset to be analyzed.}

\item{groupvar}{Group variable name. If not specified, descriptive stats for all subjects are reported.}

\item{total}{(default=TRUE). Valid only if groupvar is not NULL. total=FALSE if a user does not want to print stats for all subjects.}
}
\value{
squareroot of matrix
}
\description{
Create descriptive table by group
Current version works in the html setting. We are working on extend it to the word document
}
\examples{
data(mtcars)
mtcars = mtcars \%>\% mutate(cyl=factor(cyl),vs=factor(vs), am=factor(am))
varlist=c('mpg','disp','hp','drat','wt','gear','carb','vs')
desc_table(varlist, data=mtcars,groupvar=NULL)
desc_table(varlist, data=mtcars,groupvar="cyl")
}
\references{
TBA
}
\author{
Seonjoo Lee, \email{sl3670@cumc.columbia.edu}
}
\keyword{L1penalization}
\keyword{highdimensional}
\keyword{mediation}
