% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{Mode}
\alias{Mode}
\title{Finds the modes (modi) central tendency}
\usage{
Mode(x, out = NULL, warn = T)
}
\arguments{
\item{x}{a vector/factor for which to find the mode}

\item{out}{an integer that defines the maximum number of modi to display; a warning is shown if more modi are found}
}
\description{
Finds the modes (modi) central tendency
}
